Update games.json
I'm not sure if i did it right, but there should be a steam entry for ghost recon breakpoint, it's been released on there recently and with a few work arounds from gamingonlinux it can be played rather smoothly, albeit in a strange non full screen way. Anticheat works as well , the battleye config file must be edited and the game launched in vulkan mode. I apologize if my edit is absolute garbage but i'm not a programmer, i just wanted to have the site updated with that game.